requirements.txt
langchain==0.3.25
langchain_groq==0.3.2
python-dotenv==1.1.0
reportlab==4.4.1
streamlit==1.45.1
